requirements.txt
# requirements.txt
# Install these packages with: pip install -r requirements.txt

# Data fetching and processing
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Technical analysis
TA-Lib>=0.4.25

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Excel export
xlsxwriter>=3.0.0

# Configuration
pyyaml>=6.0

# Optimization (optional but recommended)
cvxpy>=1.3.0  # For advanced portfolio optimization
scikit-learn>=1.2.0  # For factor weight learning

# Development tools (optional)
pytest>=7.0.0
pytest-asyncio>=0.20.0
mypy>=1.0.0
black>=23.0.0

# Note for TA-Lib installation:
# On Windows: pip install TA-Lib
# On Mac: brew install ta-lib && pip install TA-Lib  
# On Linux: sudo apt-get install libta-lib-dev && pip install TA-Lib
# 
# If TA-Lib installation fails, you can use pandas-ta as an alternative:
# pip install pandas-ta
# (You'll need to modify the technical indicators in the code accordingly)